require announcements configured replaced woth require admin user whne needed

+++ control-announcements/src/main/java/org/oskari/announcements/actions/DeleteAnnouncement.java
@@ -16,7 +16,7 @@ public class DeleteAnnouncement extends AnnouncementsRestActionHandler {
 
     @Override
     public void handlePost(ActionParameters params) throws ActionException {
-        requireAnnouncementsConfigured();
+        params.requireAdminUser();
 
         try {
             JSONObject result = AnnouncementsDBHelper.deleteAnnouncement(params);
